feat(code-highlighting): expand available languages (#6867)
* feat(code-highlighting): expand standard languages

Terraform was not included because the supported terraform grammar for
highlight-js has no maintainer and no published package.

Toml is aliased to ini. See https://highlightjs.readthedocs.io/en/latest/supported-languages.html

* feat(code-highlighting): extend available languages

Adds three new exports: mediumLanguages (those less popular than
standard, but moreso than specialized), specializedLanguages, and
allLanguages.

* docs(changeset): feat: add ~45 new highlight.js languages + alias maps and new grouped exports (mediumLanguages, specializedLanguages, allLanguages)

* chore(scripts): remove container if it exists before run

* test(components): update snapshot test

The language was previously being guessed as elixir. With the expanded
language inputs it is now guessed to be properties, and as a result the
highlighting is slightly different.

* docs(components, snapshots): Update README for running tests

Adds documentation to the README for successfully running snapshot tests
in packages/components.

* doc(code-highlight): explain the language exports

Adds comments explaining the different intent of each exported language
group. For most cases standardLanguages should be sufficient. For users
looking for as close to Shiki's full language list as possible, the
allLanguages is best.